requirements.txt
-i https://pypi.org/simple
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
cryptography==41.0.2
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==2.0.0
django==4.2.3
django-ckeditor==6.6.1
django-cors-headers==4.2.0
django-js-asset==2.1.0; python_version >= '3.8'
django-tinymce==3.6.1
fonttools==4.41.0
fpdf2==2.7.4
gunicorn==21.2.0
html5lib==1.1
idna==3.4
lxml==4.9.3
oscrypto==1.3.0
packaging==23.1; python_version >= '3.7'
pillow==10.0.0
psycopg==3.1.9
psycopg2-binary==2.9.6
pycparser==2.21
pyhanko==0.19.0
pyhanko-certvalidator==0.23.0
pypdf3==1.0.6
pypng==0.20220715.0
python-bidi==0.4.2
pyyaml==6.0
qrcode==7.4.2
reportlab==4.0.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.65.0
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.3
webencodings==0.5.1
xhtml2pdf==0.2.8
